# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_in_days": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ApiKeyCreateResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "created_at",
          "key"
        ],
        "type": "object"
      },
      "ApiKeyJSON": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_used_at": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "created_at"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "yauth_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "OpenAPI 3.1 specification for the yauth-go authentication library. Routes are mounted by each plugin under whatever prefix the embedder chooses (commonly `/api/auth`).",
    "license": {
      "name": "MIT"
    },
    "title": "yauth-go",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api-keys": {
      "get": {
        "operationId": "apiKeyList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyJSON"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of keys."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not authenticated."
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "List the caller's API keys (no secrets)",
        "tags": [
          "api-key"
        ]
      },
      "post": {
        "operationId": "apiKeyCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreateRequest"
              }
            }
          },
          "description": "Name, optional scopes, optional expires_in_days",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            },
            "description": "Key created. The `secret` field is the plaintext, shown once."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing name or invalid expiry."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not authenticated."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Per-user key cap reached."
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Create a new API key (plaintext secret returned once)",
        "tags": [
          "api-key"
        ]
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "operationId": "apiKeyDelete",
        "parameters": [
          {
            "description": "API key id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Key deleted."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not authenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Key not found or not owned by caller."
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Revoke an API key the caller owns",
        "tags": [
          "api-key"
        ]
      }
    },
...
  },
...
}